+++ plugin.php
@@ -13,7 +13,7 @@
  * Plugin URI:  https://wordpress.org/plugins/wp-plugin-dependencies
  * Description: Parses 'Requires Plugins' header and information about dependencies.
  * Author: Andy Fragen, Colin Stewart, Paul Biron
- * Version: 2.0.2.1
+ * Version: 2.0.2.2
  * License: MIT
  * Network: true
  * Requires at least: 6.0
@@ -48,6 +48,18 @@ add_filter(
 	static function( $class_name ) {
 		if ( 'WP_Plugin_Install_List_Table' === $class_name ) {
 			$class_name = __NAMESPACE__ . '\PD_Install_List_Table';
+		}
+		return $class_name;
+	}
+);
+
+// Override WP_Plugins_List_Table with our own.
+require_once __DIR__ . '/wp-admin/includes/class-pd-list-table.php';
+add_filter(
+	'wp_list_table_class_name',
+	static function( $class_name ) {
+		if ( 'WP_Plugins_List_Table' === $class_name ) {
+			$class_name = __NAMESPACE__ . '\PD_List_Table';
 		}
 		return $class_name;
 	}